Fix fast refresh compatibility for messaging
