+++ frontend/src/citizen-frontend/messages/state.tsx
@@ -31,6 +31,7 @@ import {
   messageAccountQuery,
   receivedMessagesQuery
 } from './queries'
+import { isRegularThread } from './utils'
 
 export interface MessagePageState {
   messageAccount: Result<MyAccountResponse>
@@ -55,13 +56,6 @@ const defaultState: MessagePageState = {
 }
 
 export const MessageContext = createContext<MessagePageState>(defaultState)
-
-export const isRedactedThread = (
-  thread: CitizenMessageThread
-): thread is CitizenMessageThread.Redacted => thread.type === 'Redacted'
-export const isRegularThread = (
-  thread: CitizenMessageThread
-): thread is CitizenMessageThread.Regular => thread.type === 'Regular'
 
 const markMessagesReadByThreadId = (
   thread: CitizenMessageThread,
